Interactive IT Quiz
This application / quiz contains advanced questions from IT, feel free to train your brain and learn new things.
You can also collaborate and add questions into the json file. Please be careful to add the correct answer to the first place of the JSON.
This project is open-source, but if you want to use it in your own projects, please contact me and inform me.